refactor(fiststep): refactor GeneralInformationForm component
