+++ src/pages/privileges/outlets/roles/add-role/config/GeneralInformationForm/interface.tsx
